move time stuff into it's own file
for now
