+++ projects/funhouse/funhouse_mbr/fh_homesec.h
@@ -1,0 +1,7 @@
+#pragma once
+#ifndef FH_HOMESEC_H
+#define FH_HOMESEC_H
+
+#include "secrets_homesec.h"
+
+#endif /* FH_HOMESEC_H */
